Refactor: split generator services into modules
