+++ utils/image.ts
@@ -1,0 +1,10 @@
+import { promises as fs } from 'fs';
+import * as path from 'path';
+
+export async function loadImageAsBase64(imagePath: string): Promise<string> {
+  const absolutePath = path.resolve(imagePath);
+  const buffer = await fs.readFile(absolutePath);
+  const ext = path.extname(imagePath).slice(1).toLowerCase();
+  const base64 = buffer.toString('base64');
+  return `data:image/${ext};base64,${base64}`;
+}
